Expanding counter for accumulated total of all entries scanned during remset processing.
git-svn-id: https://trac.ccs.neu.edu/svn/larceny/trunk/larceny_src@6038 c9681fa6-b8ae-40c0-a6da-707a35ad5973

+++ src/Rts/Sys/gc_t.h
@@ -89,7 +89,7 @@ struct gc {
        */
 
   int stat_max_entries_remset_scan;
-  int stat_total_entries_remset_scan;
+  long long stat_total_entries_remset_scan;
   int stat_max_remset_scan;
   int stat_max_remset_scan_cpu;
   int stat_total_remset_scan;
