+++ src/Rts/Sys/stats.h
@@ -31,7 +31,7 @@ struct gclib_stats {
   int total_remset_scan_cpu;
   int remset_scan_count;
   int max_entries_remset_scan;
-  int total_entries_remset_scan;
+  long long total_entries_remset_scan;
 
   int last_ms_gc_pause;           /* Time in core gc methods in last gc. */
   int last_ms_gc_pause_cpu;
